requirements.txt
django
djangorestframework
validate_email
pytz==2019.2
